moved host zone queries
